Update to the exception system

+++ src/main/java/com/matias/springjwt/service/AuthService.java
@@ -1,9 +1,10 @@
 package com.matias.springjwt.service;
 
 import java.util.Collections;
-import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -14,12 +15,14 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import com.matias.springjwt.dto.request.LoginRequest;
-import com.matias.springjwt.dto.request.SignupRequest;
+import com.matias.springjwt.dto.request.RegisterRequest;
 import com.matias.springjwt.dto.request.TokenRefreshRequest;
-import com.matias.springjwt.dto.response.JwtResponse;
-import com.matias.springjwt.dto.response.MessageResponse;
+import com.matias.springjwt.dto.response.LoginResponse;
+import com.matias.springjwt.dto.response.RegisterResponse;
 import com.matias.springjwt.dto.response.TokenRefreshResponse;
+import com.matias.springjwt.exception.InvalidCredentialsException;
 import com.matias.springjwt.exception.TokenRefreshException;
+import com.matias.springjwt.exception.UserAlreadyExistException;
 import com.matias.springjwt.model.RefreshTokenEntity;
 import com.matias.springjwt.model.RoleEntity;
 import com.matias.springjwt.model.UserEntity;
@@ -30,6 +33,8 @@ import com.matias.springjwt.security.dto.UserDetailsImpl;
 import com.matias.springjwt.security.jwt.JwtUtils;
 import com.matias.springjwt.security.services.RefreshTokenService;
 import com.matias.springjwt.service.abs.IAuthService;
+
+import javax.persistence.EntityNotFoundException;
 
 @Service
 public class AuthService implements IAuthService {
@@ -44,7 +49,7 @@ public class AuthService implements IAuthService {
 	IRoleRepository roleRepository;
 	
 	@Autowired
-	PasswordEncoder encoder;
+	private PasswordEncoder passwordEncoder;
 	
 	@Autowired
 	JwtUtils jwtUtils;
@@ -53,14 +58,35 @@ public class AuthService implements IAuthService {
 	RefreshTokenService refreshTokenService;
 
 	@Override
-	public JwtResponse authenticateUser(LoginRequest request) {
-		Authentication authentication = authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(request.getUsername(), request.getPassword()));
+	public RegisterResponse register(RegisterRequest request) {
+		if (userRepository.existsByUsername(request.getUsername())) {
+			throw new UserAlreadyExistException("Username is already taken.");
+		}
+		if (userRepository.existsByEmail(request.getUsername())) {
+			throw new UserAlreadyExistException("Email is already taken.");
+		}
+	    RoleEntity role = roleRepository.findByName(ERole.USER)
+			.orElseThrow(() -> new EntityNotFoundException("Error: Role is not found."));
+		UserEntity user = new UserEntity();
+		BeanUtils.copyProperties(request, user);
+		user.setPassword(passwordEncoder.encode(request.getPassword()));
+		user.setRoles(Collections.singleton(role));
+		user = userRepository.save(user);
+		RegisterResponse response = new RegisterResponse();
+		BeanUtils.copyProperties(user, response);
+	    return response;
+	}
+
+	@Override
+	public LoginResponse login(LoginRequest request) {
+		Authentication authentication = authenticationManager
+			.authenticate(new UsernamePasswordAuthenticationToken(request.getUsername(), request.getPassword()));
 		SecurityContextHolder.getContext().setAuthentication(authentication);
 		UserDetailsImpl user = (UserDetailsImpl) authentication.getPrincipal();
 		String token = jwtUtils.generateJwtToken(user);
-		List<String> roles = user.getAuthorities().stream().map(GrantedAuthority::getAuthority).collect(Collectors.toList());
+		Set<String> roles = user.getAuthorities().stream().map(GrantedAuthority::getAuthority).collect(Collectors.toSet());
 		RefreshTokenEntity refreshToken = refreshTokenService.createRefreshToken(user.getId());
-		return JwtResponse.builder()
+		return LoginResponse.builder()
 				.token(token)
 				.refreshToken(refreshToken.getToken())
 				.roles(roles)
@@ -68,38 +94,24 @@ public class AuthService implements IAuthService {
 	}
 
 	@Override
-	public MessageResponse registerUser(SignupRequest request) {
-	    if (userRepository.existsByUsernameOrEmail(request.getUsername(), request.getEmail())) {
-	        return new MessageResponse("Error: Username or Email is already taken!");
-	    }
-	    RoleEntity role = roleRepository.findByName(ERole.ROLE_USER).orElseThrow(() -> new RuntimeException("Error: Role is not found."));
-	    UserEntity user = new UserEntity();
-	    user.setUsername(request.getUsername());
-	    user.setEmail(request.getEmail());
-	    user.setPassword(encoder.encode(request.getPassword()));
-	    user.setRoles(Collections.singleton(role));
-	    userRepository.save(user);
-	    return new MessageResponse("User registered successfully!");
-	}
-
-	@Override
 	public TokenRefreshResponse refreshToken(TokenRefreshRequest request) {
 	    String refreshToken = request.getRefreshToken();
-	    RefreshTokenEntity tokenEntity = refreshTokenService.findByToken(refreshToken).orElseThrow(() -> new TokenRefreshException(refreshToken, "Refresh token is not in database!"));
+	    RefreshTokenEntity tokenEntity = refreshTokenService.findByToken(refreshToken)
+			.orElseThrow(() -> new TokenRefreshException("Refresh token is not in database."));
 	    refreshTokenService.verifyExpiration(tokenEntity);
 	    String token = jwtUtils.generateTokenFromUsername(tokenEntity.getUser().getUsername());
 	    return new TokenRefreshResponse(token);
 	}
 
 	@Override
-	public MessageResponse logoutUser() {
-		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-		if (authentication.getPrincipal().equals("anonymousUser")) {
-			return new MessageResponse("No user is logged in or user is anonymous.");
-		}
-		UserDetailsImpl userDetails = (UserDetailsImpl) authentication.getPrincipal();
-		Integer userId = userDetails.getId();
-		refreshTokenService.deleteByUserId(userId);
-		return new MessageResponse("Log out successful!");
+	public void logout() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null || authentication.getPrincipal().equals("anonymousUser")) {
+            throw new InvalidCredentialsException("No user is logged in.");
+        }
+        UserDetailsImpl userDetails = (UserDetailsImpl) authentication.getPrincipal();
+        Integer userId = userDetails.getId();
+        refreshTokenService.deleteByUserId(userId);
 	}
+	
 }
